add loader to load all entry points by group
